Use setuptools in setup.py due to namespace packages interactions.
The interaction is between namespace packages, bdist_rpm and develop.
setuptools has an ugly hack (what else) to address the mismatch on the
bdist_rpm side, by not shipping __init__.py{c,o} that would conflict.
__init__.py should be identical and won't conflict.

I'd rather have this addressed on the develop side, as the hack can't
work on debian distros.

git-svn-id: https://vigilo-dev.si.c-s.fr/svn@365 b22e2e97-25c9-44ff-b637-2e5ceca36478